allow customization of permitted params
